+++ client/src/game/ui/tools/map.vue
@@ -5,7 +5,6 @@ import Tool from "@/game/ui/tools/tool.vue";
 
 import { GlobalPoint, Vector, LocalPoint } from "@/game/geom";
 import { layerManager } from "@/game/layers/manager";
-import { BaseRect } from "@/game/shapes/variants/baserect";
 import { Rect } from "@/game/shapes/variants/rect";
 import { l2g } from "@/game/units";
 import { SyncMode, InvalidationMode } from "../../../core/comm/types";
@@ -15,21 +14,34 @@ import { EventBus } from "@/game/event-bus";
 import { Shape } from "@/game/shapes/shape";
 import { ToolBasics } from "./ToolBasics";
 import { floorStore } from "@/game/layers/store";
-import { gameStore } from "../../store";
+import { DEFAULT_GRID_SIZE, gameStore } from "../../store";
 import { sendShapePositionUpdate, sendShapeSizeUpdate } from "@/game/api/emits/shape/core";
 
 @Component
 export default class MapTool extends Tool implements ToolBasics {
     name = ToolName.Map;
     active = false;
-    xCount = 3;
-    yCount = 3;
+    gridX = 3;
+    gridY = 3;
+    sizeX = 0;
+    sizeY = 0;
+
+    // used for region selection
+    factorX = 1;
+    factorY = 1;
+
     startPoint: GlobalPoint | null = null;
     rect: Rect | null = null;
-    shape: Shape | null = null;
+    shape: Rect | null = null;
     error = "";
 
-    shapeSelected = false;
+    ogRP: GlobalPoint | null = null;
+    ogW: number | null = null;
+    ogH: number | null = null;
+
+    manualDrag = true;
+    lock = true;
+    aspectRatio = 1;
 
     permittedTools_: ToolPermission[] = [
         { name: ToolName.Select, features: { enabled: [SelectFeatures.ChangeSelection] } },
@@ -42,8 +54,8 @@ export default class MapTool extends Tool implements ToolBasics {
     // Life cycle
 
     mounted(): void {
-        EventBus.$on("SelectionInfo.Shapes.Set", (shape: Shape[]) => {
-            this.shapeSelected = shape.length === 1;
+        EventBus.$on("SelectionInfo.Shapes.Set", (shapes: Shape[]) => {
+            this.setSelection(shapes);
         });
     }
 
@@ -53,47 +65,86 @@ export default class MapTool extends Tool implements ToolBasics {
 
     // End life cycle
 
-    removeRect(): void {
+    setSelection(shapes: readonly Shape[]): void {
+        if (shapes.length === 1 && this.shape === null && ["assetrect", "rect"].includes(shapes[0].type)) {
+            this.shape = shapes[0] as Rect;
+            this.ogRP = this.shape.refPoint;
+            this.ogW = this.shape.w;
+            this.ogH = this.shape.h;
+            this.aspectRatio = this.shape.w / this.shape.h;
+        } else if (shapes.length === 0) {
+            this.removeRect();
+        }
+    }
+
+    skipManualDrag(): void {
+        if (this.shape === null) return;
+
+        this.manualDrag = false;
+        this.gridX = this.shape.w / DEFAULT_GRID_SIZE;
+        this.gridY = this.shape.h / DEFAULT_GRID_SIZE;
+        this.sizeX = this.shape.w;
+        this.sizeY = this.shape.h;
+    }
+
+    removeRect(reset = true): void {
+        if (this.shape && reset) {
+            this.shape.refPoint = this.ogRP!;
+            this.shape.w = this.ogW!;
+            this.shape.h = this.ogH!;
+
+            sendShapePositionUpdate([this.shape], true);
+            sendShapeSizeUpdate({ shape: this.shape, temporary: true });
+            this.shape.invalidate(true);
+        }
         if (this.rect) {
             const layer = floorStore.currentLayer!;
             layer.removeShape(this.rect, SyncMode.NO_SYNC);
             this.rect = null;
         }
         this.permittedTools_ = [{ name: ToolName.Select, features: { enabled: [SelectFeatures.ChangeSelection] } }];
-        this.shapeSelected = false;
         this.shape = null;
         this.error = "";
+        this.manualDrag = true;
     }
 
-    apply(): void {
-        if (this.shape === null || this.rect === null) return;
-        if (!Number.isFinite(this.xCount) || !Number.isFinite(this.yCount) || this.xCount <= 0 || this.yCount <= 0) {
+    preview(temporary: boolean): void {
+        if (this.shape === null || (this.rect === null && this.manualDrag)) return;
+        if (!Number.isFinite(this.gridX) || !Number.isFinite(this.gridY) || this.gridX <= 0 || this.gridY <= 0) {
             this.error = "Input should be a positive number";
             return;
         }
 
-        const oldRefpoint = this.shape.refPoint;
-        const oldCenter = this.rect.center();
-
-        if (this.shape instanceof BaseRect) {
-            const xFactor = (this.xCount * gameStore.gridSize) / this.rect.w;
-            const yFactor = (this.yCount * gameStore.gridSize) / this.rect.h;
+        if (this.rect !== null) {
+            const xFactor = (this.gridX * gameStore.gridSize) / this.rect.w;
+            const yFactor = (this.gridY * gameStore.gridSize) / this.rect.h;
 
             this.shape.w *= xFactor;
             this.shape.h *= yFactor;
 
+            const oldRefpoint = this.shape.refPoint;
+            const oldCenter = this.rect.center();
+
             const delta = oldCenter.subtract(oldRefpoint);
             const newCenter = oldRefpoint.add(new Vector(xFactor * delta.x, yFactor * delta.y));
             this.shape.refPoint = this.shape.refPoint.add(oldCenter.subtract(newCenter));
-
-            sendShapePositionUpdate([this.shape], false);
-            sendShapeSizeUpdate({ shape: this.shape, temporary: false });
+        } else {
+            this.shape.w = this.sizeX;
+            this.shape.h = this.sizeY;
         }
-        this.removeRect();
+
+        sendShapePositionUpdate([this.shape], temporary);
+        sendShapeSizeUpdate({ shape: this.shape, temporary: temporary });
+        this.shape.invalidate(true);
+    }
+
+    apply(): void {
+        this.preview(false);
+        this.removeRect(false);
     }
 
     onSelect(): void {
-        this.shapeSelected = layerManager.getSelection().length === 1;
+        this.setSelection(layerManager.getSelection());
     }
 
     onDeselect(): void {
@@ -101,6 +152,7 @@ export default class MapTool extends Tool implements ToolBasics {
     }
 
     onDown(lp: LocalPoint): void {
+        if (!this.manualDrag) return;
         if (this.rect !== null || !layerManager.hasSelection()) return;
 
         const startPoint = l2g(lp);
@@ -116,7 +168,6 @@ export default class MapTool extends Tool implements ToolBasics {
         this.rect = new Rect(this.startPoint.clone(), 0, 0, { fillColour: "rgba(0,0,0,0)", strokeColour: "black" });
         this.rect.preventSync = true;
         layer.addShape(this.rect, SyncMode.NO_SYNC, InvalidationMode.NORMAL);
-        this.shape = layer.getSelection()[0];
         layer.setSelection(this.rect);
     }
 
@@ -158,6 +209,55 @@ export default class MapTool extends Tool implements ToolBasics {
             { name: ToolName.Select, features: { enabled: [SelectFeatures.Drag, SelectFeatures.Resize] } },
         ];
     }
+
+    toggleLock(): void {
+        this.lock = !this.lock;
+        if (this.lock) this.aspectRatio = this.shape!.w / this.shape!.h;
+    }
+
+    updateGridX(): void {
+        if (this.lock) {
+            if (this.manualDrag) {
+                this.gridY = this.gridX / (this.rect!.w / this.rect!.h);
+            } else {
+                this.gridY = this.gridX / this.aspectRatio;
+            }
+            this.sizeY = this.gridY * gameStore.gridSize;
+        }
+        this.sizeX = this.gridX * gameStore.gridSize;
+        if (!this.manualDrag && this.gridX > 0) this.preview(true);
+    }
+
+    updateGridY(): void {
+        if (this.lock) {
+            if (this.manualDrag) {
+                this.gridX = this.gridY * (this.rect!.w / this.rect!.h);
+            } else {
+                this.gridX = this.gridY * this.aspectRatio;
+            }
+            this.sizeX = this.gridX * gameStore.gridSize;
+        }
+        this.sizeY = this.gridY * gameStore.gridSize;
+        if (!this.manualDrag && this.gridY > 0) this.preview(true);
+    }
+
+    updateSizeX(): void {
+        if (this.lock) {
+            this.sizeY = this.sizeX / this.aspectRatio;
+            this.gridY = this.sizeY / gameStore.gridSize;
+        }
+        this.gridX = this.sizeX / gameStore.gridSize;
+        if (this.sizeX > 0) this.preview(true);
+    }
+
+    updateSizeY(): void {
+        if (this.lock) {
+            this.sizeX = this.sizeY * this.aspectRatio;
+            this.gridX = this.sizeX / gameStore.gridSize;
+        }
+        this.gridY = this.sizeY / gameStore.gridSize;
+        if (this.sizeY > 0) this.preview(true);
+    }
 }
 </script>
 
@@ -167,17 +267,47 @@ export default class MapTool extends Tool implements ToolBasics {
         v-if="selected"
         :style="{ '--detailRight': detailRight, '--detailArrow': detailArrow }"
     >
-        <template v-if="shapeSelected">
+        <template v-if="shape !== null">
             <div class="row">{{ error }}</div>
-            <template v-if="rect === null">{{ $t("game.ui.tools.map.drag_to_resize") }}</template>
+            <template v-if="rect === null && manualDrag === true">
+                <div id="map-selection-choice">
+                    <div>{{ $t("game.ui.tools.map.drag_to_resize") }}</div>
+                    <div id="next" @click="skipManualDrag">
+                        Scale full image instead
+                        <font-awesome-icon icon="arrow-right" />
+                    </div>
+                </div>
+            </template>
             <template v-else>
-                <div class="explanation" v-t="'game.ui.tools.map.set_target_grid_cells'"></div>
-                <div v-t="'game.ui.tools.map.horizontal'"></div>
-                <input type="number" v-model.number="xCount" class="hinput" />
-                <div v-t="'game.ui.tools.map.vertical'"></div>
-                <input type="number" v-model.number="yCount" class="vinput" />
-                <div class="button apply" @click="apply" v-t="'game.ui.tools.map.apply'"></div>
-                <div class="button cancel" @click="removeRect" v-t="'game.ui.tools.map.cancel'"></div>
+                <div id="map-grid">
+                    <div class="explanation" v-t="'game.ui.tools.map.set_target_grid_cells'"></div>
+                    <div class="map-lock" @click="toggleLock" title="(Un)lock aspect ratio">
+                        <font-awesome-icon v-show="lock" icon="link" />
+                        <font-awesome-icon v-show="!lock" icon="unlink" />
+                    </div>
+                    <label for="map-g-x" v-t="'game.ui.tools.map.horizontal'"></label>
+                    <input id="map-g-x" type="number" @input="updateGridX" v-model.number="gridX" class="hinput" />
+                    <label for="map-g-y" v-t="'game.ui.tools.map.vertical'"></label>
+                    <input id="map-g-y" type="number" @input="updateGridY" v-model.number="gridY" class="vinput" />
+                </div>
+                <div id="map-separator"></div>
+                <div id="map-size" v-show="!manualDrag">
+                    <div class="explanation">Set target pixels</div>
+
+                    <div class="map-lock" @click="toggleLock" title="(Un)lock aspect ratio">
+                        <font-awesome-icon v-show="lock" icon="link" />
+                        <font-awesome-icon v-show="!lock" icon="unlink" />
+                    </div>
+                    <label for="map-s-x" v-t="'game.ui.tools.map.horizontal'"></label>
+                    <input id="map-s-x" type="number" @input="updateSizeX" v-model.number="sizeX" class="hinput" />
+                    <label for="map-s-y" v-t="'game.ui.tools.map.vertical'"></label>
+                    <input id="map-s-y" type="number" @input="updateSizeY" v-model.number="sizeY" class="vinput" />
+                </div>
+                <div id="map-buttons">
+                    <div class="button apply" @click="apply" v-t="'game.ui.tools.map.apply'"></div>
+                    <div style="width: 25px"></div>
+                    <div class="button cancel" @click="removeRect" v-t="'game.ui.tools.map.cancel'"></div>
+                </div>
             </template>
         </template>
         <template v-else>{{ $t("game.ui.tools.map.select_shape_msg") }}</template>
@@ -188,82 +318,145 @@ export default class MapTool extends Tool implements ToolBasics {
 .map {
     display: grid;
     grid-template-areas:
-        "text text"
-        "error error"
-        "horiz hinput"
-        "verti vinput"
-        "submit cancel";
+        "error   error   error"
+        "grid    or      size"
+        "buttons buttons buttons";
 
-    > * {
-        text-align: right;
+    #map-grid {
+        grid-area: grid;
+    }
+
+    #map-size {
+        grid-area: size;
+    }
+
+    #map-grid,
+    #map-size {
+        display: grid;
+        grid-template-areas:
+            "text text text"
+            "hlabel hinput lock"
+            "vlabel vinput lock";
+
+        .explanation {
+            grid-area: text;
+            text-align: center;
+            margin-bottom: 10px;
+            padding-bottom: 5px;
+            border-bottom: 1px solid black;
+        }
+
+        .map-lock {
+            grid-area: lock;
+            display: flex;
+            flex-direction: column;
+            justify-content: center;
+            align-items: flex-end;
+            position: relative;
+            width: 20px;
+
+            &::before {
+                content: "";
+                position: absolute;
+                background-color: transparent;
+                border: solid 1px black;
+                border-left: none;
+                border-radius: 3px;
+                right: 7px;
+                top: 10px;
+                width: 30px;
+                height: 30px;
+                z-index: -1;
+            }
+
+            &::after {
+                content: "";
+                background-color: white;
+                position: absolute;
+                right: 16px;
+                width: 4px;
+                height: 32px;
+            }
+
+            svg {
+                background-color: white;
+                transform: rotate(135deg);
+            }
+        }
+    }
+
+    #map-separator {
+        grid-area: or;
+    }
+
+    #map-buttons {
+        grid-area: buttons;
+        justify-self: center;
+        align-items: center;
+        display: flex;
+
+        .button {
+            margin-top: 10px;
+            margin-bottom: 5px;
+            padding: 5px;
+            padding-right: 8px;
+            font-size: 15px;
+            font-weight: bold;
+            text-align: center;
+            font-style: italic;
+            border: solid 1px black;
+
+            &:hover {
+                font-style: normal;
+                cursor: pointer;
+                margin-top: 8px;
+                margin-bottom: 7px;
+                box-shadow: 1px 3px;
+            }
+        }
+
+        .apply {
+            font-weight: bold;
+            box-shadow: 1px 1px 0px black;
+        }
+
+        .cancel {
+            color: lightcoral;
+            border-color: lightcoral;
+            box-shadow: 1px 1px lightcoral;
+        }
+    }
+
+    #map-selection-choice {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+
+        #next {
+            margin-top: 7px;
+            padding: 7px;
+            border-radius: 7px;
+            border: solid 1px #7c253e;
+
+            &:hover {
+                color: white;
+                background-color: #7c253e;
+                cursor: pointer;
+            }
+        }
     }
 }
 
 .row {
-    grid-column: 1 / span 2;
+    grid-column: 1 / span 3;
 }
 
-.explanation {
-    grid-area: text;
-    text-align: center;
-    margin-bottom: 10px;
-    padding-bottom: 5px;
-    border-bottom: 1px solid black;
-}
-
-.horiz {
-    grid-area: horiz;
+label {
     padding: 2px;
 }
 
-.verti {
-    grid-area: verti;
-    padding: 2px;
-}
-
-.hinput,
-.vinput {
+input[type="number"] {
     width: 75px;
     padding: 2px;
-}
-
-.hinput {
-    grid-area: hinput;
-}
-
-.vinput {
-    grid-area: vinput;
-}
-
-.button {
-    margin-top: 10px;
-    margin-bottom: 5px;
-    padding: 5px;
-    font-size: 15px;
-    font-weight: bold;
-    text-align: center;
-    font-style: italic;
-    border: solid 1px black;
-
-    &:hover {
-        font-style: normal;
-        cursor: pointer;
-        margin-top: 8px;
-        margin-bottom: 7px;
-        box-shadow: 1px 3px;
-    }
-}
-
-.apply {
-    grid-area: submit;
-    font-weight: bold;
-    box-shadow: 1px 1px 0px black;
-}
-
-.cancel {
-    grid-area: cancel;
-    color: lightcoral;
-    border-color: lightcoral;
-    box-shadow: 1px 1px lightcoral;
 }
 </style>